+++ packages/revolution/src/governance/RevolutionDAOLogicV1.sol
@@ -76,8 +76,14 @@ contract RevolutionDAOLogicV1 is
     ///                         IMMUTABLES                       ///
     ///                                                          ///
 
-    /// @notice The name of this contract
+    /// @notice The name of this revolution
     string public name;
+
+    /// @notice The purpose of this revolution
+    string public purpose;
+
+    /// @notice The flag of this revolution
+    string public flag;
 
     /// @notice The minimum settable proposal threshold
     uint256 public constant MIN_PROPOSAL_THRESHOLD_BPS = 1; // 1 basis point or 0.01%
@@ -174,7 +180,7 @@ contract RevolutionDAOLogicV1 is
         ) revert INVALID_PROPOSAL_THRESHOLD_BPS();
 
         // Initialize EIP-712 support
-        __EIP712_init(_govParams.daoName, "1");
+        __EIP712_init(_govParams.name, "1");
 
         // Grant ownership to the executor
         __Ownable_init(_executor);
@@ -189,7 +195,9 @@ contract RevolutionDAOLogicV1 is
         votingPeriod = _govParams.votingPeriod;
         votingDelay = _govParams.votingDelay;
         proposalThresholdBPS = _govParams.proposalThresholdBPS;
-        name = _govParams.daoName;
+        name = _govParams.name;
+        purpose = _govParams.purpose;
+        flag = _govParams.flag;
 
         _setDynamicQuorumParams(
             _govParams.dynamicQuorumParams.minQuorumVotesBPS,